==> notebooks/beerslaw/2 - post test/0 - Raw JSON to Dataframe.ipynb <==
import os
import re
import yaml
import json
import pickle

import numpy as np
import pandas as pd

from matplotlib import pyplot as plt
import seaborn as sns

# Data

root = '../../data/temp/'

sessions = {
    'Session 1': 1,
    'Session 2': 1,
    'Session 3': 2,
    'Session 4': 2,
    'Session 5': 2,
    'Session 6': 2,
    'Session 7': 2,
    'Session 8': 2,
    'Session 9': 2,
    'Session 10': 2,
    'Session 11': 2,
    'Session 12': 2,
    'Session 13': 3,
    'Session 14': 3,
    'Session 15': 3,
    'Session 16': 3,
    'Session 17': 3,
    'Session 18': 3,
    'Session 19': 3,
    'Session 20': 3,
    'Session 21': 3,
    'Session 22': 3,
    'Session 23': 3,
    'Session 24': 3,
    'Session 25': 3,
    'Session 26': 3,
    'Session 27': 3,
    'Session 28': 3,
    'Session 29': 3,
    'Session 30': 3,
    'Session 31': 3,
    'Session 32': 3,
    'Session 33': 3,
    'Session 34': 3,
    'Session 35': 3,
    'Session 36': 3,
    'Session 37': 3,
    'Session 38': 4,
    'Session 39': 4,
    'Session 40': 4,
    'Session 41': 4
}

# Functions

def get_data(sesh, index, field, status):
    if index in sesh:
        status = -1
        if field != '':
            if field in sesh[index]:
                return sesh[index][field], status
            else:
                return 'wrong field', status
        else:
            return sesh[index], status
    else:
        if status == -1:
            status = index
        return 'missing', status

# Version Extractions

## func

session = 'Session 1'
with open(root + session + '/cdq4drkk-q.json', 'r') as fp:
    sesh1 = json.load(fp)

def extract_version1(sesh):
    row = []
    username, status = get_data(sesh, 'name', '', -1)
    row.append(username)
    
    start_time = ''
    exploration_time = ''
    ranking_time = ''
    row.append(start_time)
    row.append(exploration_time)
    row.append(ranking_time)
    
    ranking, status = get_data(sesh, '0', 'ranks', status)
    if ranking != 'missing':
        ranking = [str(x) for x in ranking]
        ranking = ''.join(ranking)
    ranking_confidence, status = get_data(sesh, '0', 'conf', status)
    ranking_time, status = get_data(sesh, '0', 'time', status)
    row.append(ranking)
    row.append(ranking_confidence)
    row.append(ranking_time)
    
    q1, status = get_data(sesh, '1', 'text', status)
    q1_confidence, status = get_data(sesh, '1', 'conf', status)
    q1_time, status = get_data(sesh, '1', 'time', status)
    row.append(q1)
    row.append(q1_confidence)
    row.append(q1_time)

    q2, status = get_data(sesh, '2', 'text', status)
    q2_conf, status = get_data(sesh, '2', 'conf', status)
    q2_time, status = get_data(sesh, '2', 'time', status)
    row.append(q2)
    row.append(q2_conf)
    row.append(q2_time)
    
    q3, status = get_data(sesh, '3', 'text', status)
    q3_conf, status = get_data(sesh, '3', 'conf', status)
    q3_time, status = get_data(sesh, '3', 'time', status)
    row.append(q3)
    row.append(q3_conf)
    row.append(q3_time)
    
    q4, status = get_data(sesh, '4', 'text', status)
    q4_conf, status = get_data(sesh, '4', 'conf', status)
    q4_time, status = get_data(sesh, '4', 'time', status)
    row.append(q4)
    row.append(q4_conf)
    row.append(q4_time)

    q5_colours, status = get_data(sesh, '5', 'sliders', status)
    if q5_colours != 'missing':
        q5_colour0 = q5_colours[0]
        q5_colour1 = q5_colours[1]
        q5_colour2 = q5_colours[2]
        q5_colour3 = q5_colours[3]
    else:
        q5_colour0 = 'missing'
        q5_colour1 = 'missing'
        q5_colour2 = 'missing'
        q5_colour3 = 'missing'
            
    q5_time, status = get_data(sesh, '5', 'time', status)
    row.append(q5_colour0)
    row.append(q5_colour1)
    row.append(q5_colour2)
    row.append(q5_colour3)
    row.append(q5_time)

    q6_colours, status = get_data(sesh, '6', 'sliders', status)
    if q6_colours != 'missing':
        q6_colour0 = q6_colours[0]
        q6_colour1 = q6_colours[1]
        q6_colour2 = q6_colours[2]
        q6_colour3 = q6_colours[3]
    else:
        q6_colour0 = 'missing'
        q6_colour1 = 'missing'
        q6_colour2 = 'missing'
        q6_colour3 = 'missing'
    q6_time, status = get_data(sesh, '6', 'time', status)
    row.append(q6_colour0)
    row.append(q6_colour1)
    row.append(q6_colour2)
    row.append(q6_colour3)
    row.append(q6_time)
    
    row = row + ['missing' for _ in range(10)]
    
    formula, status = get_data(sesh, '7', 'text', status)
    formula_conf, status = get_data(sesh, '7', 'conf', status)
    formula_time, status = get_data(sesh, '7', 'time', status)
    row.append(formula)
    row.append(formula_conf)
    row.append(formula_time)

    notes_text = ''
    notes_math = ''
    notes_table = ''
    notes_diag = ''
    notes_time = ''
    row.append(notes_text)
    row.append(notes_math)
    row.append(notes_table)
    row.append(notes_diag)
    row.append(notes_time)

    beerslaw, status = get_data(sesh, '8', 'text', status)
    beerslaw_conf, status = get_data(sesh, '8', 'conf', status)
    beerslaw_time, status = get_data(sesh, '8', 'time', status)
    row.append(beerslaw)
    row.append(beerslaw_conf)
    row.append(beerslaw_time)

    problem_ranking = ''
    problem_conf = ''
    problem_time = ''
    row.append(problem_ranking)
    row.append(problem_conf)
    row.append(problem_time)

    feedback_entertain, status = get_data(sesh, 'feedback', 's_entertain', status)
    feedback_difficult, status = get_data(sesh, 'feedback', 's_difficult', status)
    feedback_good, status = get_data(sesh, 'feedback', 't_good', status)
    feedback_bad, status = get_data(sesh, 'feedback', 't_bad', status)
    feedback, status = get_data(sesh, 'feedback', 'status', status)
    feedback_time, status = get_data(sesh, 'feedback', 'time', status)
    row.append(feedback_entertain)
    row.append(feedback_difficult)
    row.append(feedback_good)
    row.append(feedback_bad)
    row.append(feedback)
    row.append(feedback_time)

    experiment_time = ''
    row.append(experiment_time)
    
    
    if status == -1:
        status = 'complete'
    row.append(status)

    return row

len(extract_version1(sesh1))

## func

def extract_version2(sesh):
    row = []
    username, status = get_data(sesh, 'name', '', -1)
    row.append(username)
    
    start_time, status = get_data(sesh, '1', '', status)
    exploration_time, status = get_data(sesh, '2', '', status)
    ranking_time, status = get_data(sesh, '3', '', status)
    row.append(start_time)
    row.append(exploration_time)
    row.append(ranking_time)
    
    ranking, status = get_data(sesh, '4', 'ranks', status)
    if ranking != 'missing':
        ranking = [str(x) for x in ranking]
        ranking = ''.join(ranking)
    ranking_confidence, status = get_data(sesh, '4', 'conf', status)
    ranking_time, status = get_data(sesh, '4', 'time', status)
    row.append(ranking)
    row.append(ranking_confidence)
    row.append(ranking_time)
    
    q1, status = get_data(sesh, '5', 'text', status)
    q1_confidence, status = get_data(sesh, '5', 'conf', status)
    q1_time, status = get_data(sesh, '5', 'time', status)
    row.append(q1)
    row.append(q1_confidence)
    row.append(q1_time)

    q2, status = get_data(sesh, '6', 'text', status)
    q2_conf, status = get_data(sesh, '6', 'conf', status)
    q2_time, status = get_data(sesh, '6', 'time', status)
    row.append(q2)
    row.append(q2_conf)
    row.append(q2_time)
    
    q3, status = get_data(sesh, '7', 'text', status)
    q3_conf, status = get_data(sesh, '7', 'conf', status)
    q3_time, status = get_data(sesh, '7', 'time', status)
    row.append(q3)
    row.append(q3_conf)
    row.append(q3_time)
    
    q4, status = get_data(sesh, '8', 'text', status)
    q4_conf, status = get_data(sesh, '8', 'conf', status)
    q4_time, status = get_data(sesh, '8', 'time', status)
    row.append(q4)
    row.append(q4_conf)
    row.append(q4_time)

    q5_colours, status = get_data(sesh, '9', 'sliders', status)
    if q5_colours != 'missing':
        q5_colour0 = q5_colours[0]
        q5_colour1 = q5_colours[1]
        q5_colour2 = q5_colours[2]
        q5_colour3 = q5_colours[3]
    else:
        q5_colour0 = 'missing'
        q5_colour1 = 'missing'
        q5_colour2 = 'missing'
        q5_colour3 = 'missing'
            
    q5_time, status = get_data(sesh, '9', 'time', status)
    row.append(q5_colour0)
    row.append(q5_colour1)
    row.append(q5_colour2)
    row.append(q5_colour3)
    row.append(q5_time)

    q6_colours, status = get_data(sesh, '10', 'sliders', status)
    if q6_colours != 'missing':
        q6_colour0 = q6_colours[0]
        q6_colour1 = q6_colours[1]
        q6_colour2 = q6_colours[2]
        q6_colour3 = q6_colours[3]
    else:
        q6_colour0 = 'missing'
        q6_colour1 = 'missing'
        q6_colour2 = 'missing'
        q6_colour3 = 'missing'
    q6_time, status = get_data(sesh, '10', 'time', status)
    row.append(q6_colour0)
    row.append(q6_colour1)
    row.append(q6_colour2)
    row.append(q6_colour3)
    row.append(q6_time)
    
    row = row + ['missing' for _ in range(10)]
    
    formula, status = get_data(sesh, '11', 'text', status)
    formula_conf, status = get_data(sesh, '11', 'conf', status)
    formula_time, status = get_data(sesh, '11', 'time', status)
    row.append(formula)
    row.append(formula_conf)
    row.append(formula_time)

    notes, status = get_data(sesh, '12', 'choices', status)
    if notes != 'missing':
        notes_text = notes[0]
        notes_math = notes[1]
        notes_table = notes[2]
        notes_diag = notes[3]
    else:
        notes_text = ''
        notes_math = ''
        notes_table = ''
        notes_diag = ''
    notes_time, status = get_data(sesh, '12', 'time', status)
    row.append(notes_text)
    row.append(notes_math)
    row.append(notes_table)
    row.append(notes_diag)
    row.append(notes_time)

    beerslaw, status = get_data(sesh, '13', 'text', status)
    beerslaw_conf, status = get_data(sesh, '13', 'conf', status)
    beerslaw_time, status = get_data(sesh, '13', 'time', status)
    row.append(beerslaw)
    row.append(beerslaw_conf)
    row.append(beerslaw_time)

    problem_ranking, status = get_data(sesh, '15', 'choices', status)
    problem_conf, status = get_data(sesh, '15', 'confidence', status)
    problem_time, status = get_data(sesh, '15', 'time', status)
    row.append(problem_ranking)
    row.append(problem_conf)
    row.append(problem_time)

    feedback_entertain, status = get_data(sesh, 'feedback', 's_entertain', status)
    feedback_difficult, status = get_data(sesh, 'feedback', 's_difficult', status)
    feedback_good, status = get_data(sesh, 'feedback', 't_good', status)
    feedback_bad, status = get_data(sesh, 'feedback', 't_bad', status)
    feedback, status = get_data(sesh, 'feedback', 'status', status)
    feedback_time, status = get_data(sesh, 'feedback', 'time', status)
    row.append(feedback_entertain)
    row.append(feedback_difficult)
    row.append(feedback_good)
    row.append(feedback_bad)
    row.append(feedback)
    row.append(feedback_time)

    experiment_time, status = get_data(sesh, '17', 'time', status)
    row.append(experiment_time)
    
    
    if status == -1:
        status = 'complete'
    row.append(status)

    return row

session = 'Session 3'
with open(root + session + '/4vtjgtja-q.json', 'r') as fp:
    sesh3 = json.load(fp)

len(extract_version2(sesh3))

## Func

def extract_version3(sesh):
    row = []
    username, status = get_data(sesh, 'name', '', -1)
    row.append(username)
    
    start_time, status = get_data(sesh, '1', '', status)
    exploration_time, status = get_data(sesh, '2', '', status)
    ranking_time, status = get_data(sesh, '3', '', status)
    row.append(start_time)
    row.append(exploration_time)
    row.append(ranking_time)
    
    ranking, status = get_data(sesh, '4', 'choices', status)
    if ranking != 'missing':
        ranking = [str(x) for x in ranking]
        ranking = ''.join(ranking)
        ranking = ranking.replace('Becher A', '0')
        ranking = ranking.replace('Becher B', '1')
        ranking = ranking.replace('Becher C', '2')
        ranking = ranking.replace('Becher D', '3')
    ranking_confidence, status = get_data(sesh, '4', 'conf', status)
    ranking_time, status = get_data(sesh, '4', 'time', status)
    row.append(ranking)
    row.append(ranking_confidence)
    row.append(ranking_time)
    
    q1, status = get_data(sesh, '5', 'text', status)
    q1_confidence, status = get_data(sesh, '5', 'conf', status)
    q1_time, status = get_data(sesh, '5', 'time', status)
    row.append(q1)
    row.append(q1_confidence)
    row.append(q1_time)

    q2, status = get_data(sesh, '6', 'text', status)
    q2_conf, status = get_data(sesh, '6', 'conf', status)
    q2_time, status = get_data(sesh, '6', 'time', status)
    row.append(q2)
    row.append(q2_conf)
    row.append(q2_time)
    
    q3, status = get_data(sesh, '7', 'text', status)
    q3_conf, status = get_data(sesh, '7', 'conf', status)
    q3_time, status = get_data(sesh, '7', 'time', status)
    row.append(q3)
    row.append(q3_conf)
    row.append(q3_time)
    
    q4, status = get_data(sesh, '8', 'text', status)
    q4_conf, status = get_data(sesh, '8', 'conf', status)
    q4_time, status = get_data(sesh, '8', 'time', status)
    row.append(q4)
    row.append(q4_conf)
    row.append(q4_time)

    q5_colours, status = get_data(sesh, '9', 'sliders', status)
    if q5_colours != 'missing':
        q5_colour0 = q5_colours[0]
        q5_colour1 = q5_colours[1]
        q5_colour2 = q5_colours[2]
        q5_colour3 = q5_colours[3]
    else:
        q5_colour0 = 'missing'
        q5_colour1 = 'missing'
        q5_colour2 = 'missing'
        q5_colour3 = 'missing'
            
    q5_time, status = get_data(sesh, '9', 'time', status)
    row.append(q5_colour0)
    row.append(q5_colour1)
    row.append(q5_colour2)
    row.append(q5_colour3)
    row.append(q5_time)

    q6_colours, status = get_data(sesh, '10', 'sliders', status)
    if q6_colours != 'missing':
        q6_colour0 = q6_colours[0]
        q6_colour1 = q6_colours[1]
        q6_colour2 = q6_colours[2]
        q6_colour3 = q6_colours[3]
    else:
        q6_colour0 = 'missing'
        q6_colour1 = 'missing'
        q6_colour2 = 'missing'
        q6_colour3 = 'missing'
    q6_time, status = get_data(sesh, '10', 'time', status)
    row.append(q6_colour0)
    row.append(q6_colour1)
    row.append(q6_colour2)
    row.append(q6_colour3)
    row.append(q6_time)
    
    row = row + ['missing' for _ in range(10)]
    
    formula, status = get_data(sesh, '11', 'text', status)
    formula_conf, status = get_data(sesh, '11', 'conf', status)
    formula_time, status = get_data(sesh, '11', 'time', status)
    row.append(formula)
    row.append(formula_conf)
    row.append(formula_time)

    notes, status = get_data(sesh, '12', 'choices', status)
    if notes != 'missing':
        notes_text = notes[0]
        notes_math = notes[1]
        notes_table = notes[2]
        notes_diag = notes[3]
    else:
        notes_text = ''
        notes_math = ''
        notes_table = ''
        notes_diag = ''
    notes_time, status = get_data(sesh, '12', 'time', status)
    row.append(notes_text)
    row.append(notes_math)
    row.append(notes_table)
    row.append(notes_diag)
    row.append(notes_time)

    beerslaw, status = get_data(sesh, '13', 'text', status)
    beerslaw_conf, status = get_data(sesh, '13', 'conf', status)
    beerslaw_time, status = get_data(sesh, '13', 'time', status)
    row.append(beerslaw)
    row.append(beerslaw_conf)
    row.append(beerslaw_time)

    problem_ranking, status = get_data(sesh, '15', 'choices', status)
    problem_conf, status = get_data(sesh, '15', 'confidence', status)
    problem_time, status = get_data(sesh, '15', 'time', status)
    row.append(problem_ranking)
    row.append(problem_conf)
    row.append(problem_time)

    feedback_entertain, status = get_data(sesh, 'feedback', 's_entertain', status)
    feedback_difficult, status = get_data(sesh, 'feedback', 's_difficult', status)
    feedback_good, status = get_data(sesh, 'feedback', 't_good', status)
    feedback_bad, status = get_data(sesh, 'feedback', 't_bad', status)
    feedback, status = get_data(sesh, 'feedback', 'status', status)
    feedback_time, status = get_data(sesh, 'feedback', 'time', status)
    row.append(feedback_entertain)
    row.append(feedback_difficult)
    row.append(feedback_good)
    row.append(feedback_bad)
    row.append(feedback)
    row.append(feedback_time)

    experiment_time, status = get_data(sesh, '17', 'time', status)
    row.append(experiment_time)
    
    
    if status == -1:
        status = 'complete'
    row.append(status)

    return row

session = 'Session 13'
with open(root + session + '/4aaspwgs-q.json', 'r') as fp:
    sesh13 = json.load(fp)

len(extract_version3(sesh13))

## func

session = 'Session 40'
with open(root + session + '/9wdm7zue-q.json', 'r') as fp:
    sesh40 = json.load(fp)

def extract_version4(sesh):
    row = []
    username, status = get_data(sesh, 'name', '', -1)
    row.append(username)
    
    start_time, status = get_data(sesh, '1', '', status)
    exploration_time, status = get_data(sesh, '2', '', status)
    ranking_time, status = get_data(sesh, '3', '', status)
    row.append(start_time)
    row.append(exploration_time)
    row.append(ranking_time)
    
    ranking, status = get_data(sesh, '4', 'choices', status)
    if ranking != 'missing':
        ranking = [str(x) for x in ranking]
        ranking = ''.join(ranking)
        ranking = ranking.replace('Becher A', '0')
        ranking = ranking.replace('Becher B', '1')
        ranking = ranking.replace('Becher C', '2')
        ranking = ranking.replace('Becher D', '3')
    ranking_confidence, status = get_data(sesh, '4', 'conf', status)
    ranking_time, status = get_data(sesh, '4', 'time', status)
    row.append(ranking)
    row.append(ranking_confidence)
    row.append(ranking_time)
    
    q1, status = get_data(sesh, '5', 'text', status)
    q1_confidence, status = get_data(sesh, '5', 'conf', status)
    q1_time, status = get_data(sesh, '5', 'time', status)
    row.append(q1)
    row.append(q1_confidence)
    row.append(q1_time)

    q2, status = get_data(sesh, '6', 'text', status)
    q2_conf, status = get_data(sesh, '6', 'conf', status)
    q2_time, status = get_data(sesh, '6', 'time', status)
    row.append(q2)
    row.append(q2_conf)
    row.append(q2_time)
    
    q3, status = get_data(sesh, '7', 'text', status)
    q3_conf, status = get_data(sesh, '7', 'conf', status)
    q3_time, status = get_data(sesh, '7', 'time', status)
    row.append(q3)
    row.append(q3_conf)
    row.append(q3_time)
    
    q4, status = get_data(sesh, '8', 'text', status)
    q4_conf, status = get_data(sesh, '8', 'conf', status)
    q4_time, status = get_data(sesh, '8', 'time', status)
    row.append(q4)
    row.append(q4_conf)
    row.append(q4_time)

    q5_colours, status = get_data(sesh, '11', 'sliders', status)
    if q5_colours != 'missing':
        q5_colour0 = q5_colours[0]
        q5_colour1 = q5_colours[1]
        q5_colour2 = q5_colours[2]
        q5_colour3 = q5_colours[3]
    else:
        q5_colour0 = 'missing'
        q5_colour1 = 'missing'
        q5_colour2 = 'missing'
        q5_colour3 = 'missing'
            
    q5_time, status = get_data(sesh, '11', 'time', status)
    row.append(q5_colour0)
    row.append(q5_colour1)
    row.append(q5_colour2)
    row.append(q5_colour3)
    row.append(q5_time)

    q6_colours, status = get_data(sesh, '12', 'sliders', status)
    if q6_colours != 'missing':
        q6_colour0 = q6_colours[0]
        q6_colour1 = q6_colours[1]
        q6_colour2 = q6_colours[2]
        q6_colour3 = q6_colours[3]
    else:
        q6_colour0 = 'missing'
        q6_colour1 = 'missing'
        q6_colour2 = 'missing'
        q6_colour3 = 'missing'
    q6_time, status = get_data(sesh, '12', 'time', status)
    row.append(q6_colour0)
    row.append(q6_colour1)
    row.append(q6_colour2)
    row.append(q6_colour3)
    row.append(q6_time)
    
    q7_colours, status = get_data(sesh, '9', 'sliders', status)
    if q7_colours != 'missing':
        q7_colour0 = q7_colours[0]
        q7_colour1 = q7_colours[1]
        q7_colour2 = q7_colours[2]
        q7_colour3 = q7_colours[3]
    else:
        q7_colour0 = 'missing'
        q7_colour1 = 'missing'
        q7_colour2 = 'missing'
        q7_colour3 = 'missing'
    q7_time, status = get_data(sesh, '9', 'time', status)
    row.append(q7_colour0)
    row.append(q7_colour1)
    row.append(q7_colour2)
    row.append(q7_colour3)
    row.append(q7_time)
    
    q8_colours, status = get_data(sesh, '10', 'sliders', status)
    if q8_colours != 'missing':
        q8_colour0 = q8_colours[0]
        q8_colour1 = q8_colours[1]
        q8_colour2 = q8_colours[2]
        q8_colour3 = q8_colours[3]
    else:
        q8_colour0 = 'missing'
        q8_colour1 = 'missing'
        q8_colour2 = 'missing'
        q8_colour3 = 'missing'
    q8_time, status = get_data(sesh, '10', 'time', status)
    row.append(q8_colour0)
    row.append(q8_colour1)
    row.append(q8_colour2)
    row.append(q8_colour3)
    row.append(q8_time)
    
    formula, status = get_data(sesh, '13', 'text', status)
    formula_conf, status = get_data(sesh, '13', 'conf', status)
    formula_time, status = get_data(sesh, '13', 'time', status)
    row.append(formula)
    row.append(formula_conf)
    row.append(formula_time)

    notes, status = get_data(sesh, '14', 'choices', status)
    if notes != 'missing':
        notes_text = notes[0]
        notes_math = notes[1]
        notes_table = notes[2]
        notes_diag = notes[3]
    else:
        notes_text = ''
        notes_math = ''
        notes_table = ''
        notes_diag = ''
    notes_time, status = get_data(sesh, '14', 'time', status)
    row.append(notes_text)
    row.append(notes_math)
    row.append(notes_table)
    row.append(notes_diag)
    row.append(notes_time)

    beerslaw, status = get_data(sesh, '15', 'text', status)
    beerslaw_conf, status = get_data(sesh, '15', 'conf', status)
    beerslaw_time, status = get_data(sesh, '15', 'time', status)
    row.append(beerslaw)
    row.append(beerslaw_conf)
    row.append(beerslaw_time)

    problem_ranking, status = get_data(sesh, '17', 'choices', status)
    problem_conf, status = get_data(sesh, '17', 'confidence', status)
    problem_time, status = get_data(sesh, '17', 'time', status)
    row.append(problem_ranking)
    row.append(problem_conf)
    row.append(problem_time)

    feedback_entertain, status = get_data(sesh, 'feedback', 's_entertain', status)
    feedback_difficult, status = get_data(sesh, 'feedback', 's_difficult', status)
    feedback_good, status = get_data(sesh, 'feedback', 't_good', status)
    feedback_bad, status = get_data(sesh, 'feedback', 't_bad', status)
    feedback, status = get_data(sesh, 'feedback', 'status', status)
    feedback_time, status = get_data(sesh, 'feedback', 'time', status)
    row.append(feedback_entertain)
    row.append(feedback_difficult)
    row.append(feedback_good)
    row.append(feedback_bad)
    row.append(feedback)
    row.append(feedback_time)

    experiment_time, status = get_data(sesh, '19', 'time', status)
    row.append(experiment_time)
    
    
    if status == -1:
        status = 'complete'
    row.append(status)

    return row

len(extract_version4(sesh40))

# Parsing

# Version to  Version Functions
version_function = {
    1: extract_version1,
    2: extract_version2,
    3: extract_version3,
    4: extract_version4
}

# get the session number
def get_session_number(path: str):
    reg = '(Session [0-9]+)'
    reg = re.compile(reg)
    reg = reg.findall(path)
    if len(reg) == 0:
        print(path, reg)
        {'hello': 0}['bye']
    return reg[0]

# reads the post test files and extract the information
def process_version(path: str, sessions: dict, version_function: dict):
    try:
        session = get_session_number(path)
        version = sessions[session]
        fun = version_function[version]
        with open(path, 'r') as fp:
            sesh = json.load(fp)
        row = fun(sesh)
        row.append(session)
        row.append(version)
    except KeyError as ke:
        return []
    return row
        
    

# Crawl all of the files
repo = '../../data/temp/'
files = []
for r, d, f in os.walk(repo):
    for file in f:
        if file.endswith(".json") and 'prefills.json' not in file:
            files.append(os.path.join(r, file))

# Pull the information and create the dataframe
post_test = []
for file in files:
    row = process_version(file, sessions, version_function)
    post_test.append(row)
    
# Columns
columns = [
    'username',
    'start_time', 'exploration_time', 'ranking_task_time',
    'ranking', 'ranking_confidence', 'ranking_time',
    'q1', 'q1_conf', 'q1_time',
    'q2', 'q2_conf', 'q2_time',
    'q3', 'q3_conf', 'q3_time',
    'q4', 'q4_conf', 'q4_time',
    'q5_colour0', 'q5_colour1', 'q5_colour2', 'q5_colour3', 'q5_time',
    'q6_colour0', 'q6_colour1', 'q6_colour2', 'q6_colour3', 'q6_time',
    'q7_colour0', 'q7_colour1', 'q7_colour2', 'q7_colour3', 'q7_time',
    'q8_colour0', 'q8_colour1', 'q8_colour2', 'q8_colour3', 'q8_time',
    'formula', 'fomula_conf', 'formula_time',
    'notes_text', 'notes_math', 'notes_table', 'notes_diag', 'notes_time',
    'beerslaw', 'beerslaw_conf', 'beerslaw_time',
    'problem_ranking', 'problem_conf', 'problem_time', 
    'feedback_entertain', 'feedback_difficult', 'feedback_good', 'feedback_bad', 'feedback', 'feedback_time',
    'experiment_time', 'status', 'session', 'version'
]

# DataFrame
post_test = pd.DataFrame(post_test, columns=columns)

post_test

post_test.to_csv('../../data/post_test/extracted.csv', sep='\t')
with open('../../data/post_test/extracted.pkl', 'wb') as fp:
    pickle.dump(post_test, fp)
