==> notebooks/beerslaw/2 - post test/2 - Merge all data.ipynb <==
This notebook is used to bring the data about the sessions, schools, field, demographics, ... together!

import os
import re
import pickle

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)

# Data

schedule = pd.read_csv('../../data/post_test/schedule.csv')
schedule['session_nr'] = schedule['Session'].apply(lambda x: re.findall(r'\d+', x)[0])
schedule_columns = [
    'Language', 'Field of Study', 'Study Level', 'Year', 'Session'
]

schedule = schedule[schedule_columns]
schedule.columns = [
    'language', 'field', 'level', 'year', 'session'
]

users = pd.read_csv('../../data/post_test/users.csv')
users_columns = [
    'username', 'consent', 'gender'
]
users = users[users_columns]

with open('../../data/post_test/questions.pkl', 'rb') as fp:
    questions = pickle.load(fp)
    
with open('../../data/post_test/rankings.pkl', 'rb') as fp:
    rankings = pickle.load(fp)
    
with open('../../data/post_test/some_questions.pkl', 'rb') as fp:
    some_questions = pickle.load(fp)
    
with open('../../data/post_test/some.pkl', 'rb') as fp:
    some = pickle.load(fp)

# Merging of the data

print(len(questions))
questions = questions.merge(schedule, on='session', how='left')
questions = questions.merge(users, on='username', how='left')
print(len(questions))

print(len(rankings))
rankings = rankings.merge(schedule, on='session', how='left')
rankings = rankings.merge(users, on='username', how='left')
print(len(rankings))

print(len(some_questions))
some_questions = some_questions.merge(schedule, on='session', how='left')
some_questions = some_questions.merge(users, on='username', how='left')
print(len(some_questions))

print(len(some))
some = some.merge(schedule, on='session', how='left')
some = some.merge(users, on='username', how='left')
print(len(some))

with open('../../data/post_test/questions.pkl', 'wb') as fp:
    pickle.dump(questions, fp)
    # full questions
    
with open('../../data/post_test/rankings.pkl', 'wb') as fp:
    pickle.dump(rankings, fp)
    # rankings completed
    
with open('../../data/post_test/some_questions.pkl', 'wb') as fp:
    pickle.dump(some_questions, fp)
    # partial questions
    
with open('../../data/post_test/some.pkl', 'wb') as fp:
    pickle.dump(some, fp)
    # partial questions
