==> notebooks/beerslaw/2 - post test/3 - Translating.ipynb <==
import os
import pickle

import numpy as np
import pandas as pd

from googletrans import Translator

# Data

with open('../../data/post_test/questions.pkl', 'rb') as fp:
    questions = pickle.load(fp)
    # full questions
    
with open('../../data/post_test/rankings.pkl', 'rb') as fp:
    rankings = pickle.load(fp)
    # rankings completed
    
with open('../../data/post_test/some_questions.pkl', 'rb') as fp:
    some_questions = pickle.load(fp)
    # partial questions

print(questions.columns)
print(rankings.columns)
print(some_questions.columns)

# Translate

translator = Translator()

def translate(text:str) -> str:
    try:
        translate = translator.translate(text, src='de', dest='en').text
    except TypeError:
        translate = 'type error'
    return translate
translator.raise_Exception = True

questions['beerslaw_en'] = questions['beerslaw'].apply(translate)
questions['formula_en'] = questions['formula'].apply(translate)

rankings['beerslaw_en'] = rankings['beerslaw'].apply(lambda x: translate)
rankings['formula_en'] = rankings['formula'].apply(lambda x: translate)

some_questions['beerslaw_en'] = some_questions['beerslaw'].apply(translate)
some_questions['formula_en'] = some_questions['formula'].apply(translate)

with open('../../data/post_test/questions.pkl', 'wb') as fp:
    pickle.dump(questions, fp)
    # full questions
    
with open('../../data/post_test/rankings.pkl', 'wb') as fp:
    pickle.dump(rankings, fp)
    # rankings completed
    
with open('../../data/post_test/some_questions.pkl', 'wb') as fp:
    pickle.dump(some_questions, fp)
    # partial questions


results =translator.translate('Je suis un petit canard')
print(results.text)
